Fix random doxygen warnings and typos
Silence annoying
"/builds/gromacs/gromacs/src/gromacs/math/vectypes.h:87: error:
gmx::BasicVector: is in library file(s), but appears in public
documentation" from source-check pipeline (appears to be false-positive
due to a forward-declaration in a random file). I was not able to come
up with a quick fix for the underlying problem in check-source, and,
from my understanding, there are plans for refactoring it soon anyway;
so I went with a suppression here.

Fix a bunch of random typos throughout the code.

Fix some unbalanced Doxygen's \{ and \}. And one widowed @}.

Fix some \copydocs (not all).

Remove a nasty non-breakable-space that was messing up block
indentation in Sphinx docs.