Halo exchange unit test for pre-existing CPU codepath
The test sets up a 2D rank topology and performs a coordinate halo
exchange (using the pre-existing CPU codepath), with 2 pulses in
the first dimension and 1 pulse in the second. Each pulse involves
a few non-contiguous indices. The sending rank, atom number and
spatial 3D index are encoded in the x values, to allow correctness
checking following the halo exchange. Follow-up will port to GPU codepath.